Backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
flask-socketio==5.1.1
Werkzeug==2.0.3
flask_pymongo==2.3.0
dnspython
redis==4.2.0